Python3 code compatibility: division, round
Make division transparent-compatible (from __future__ import division, // for strict int result)
Do not use round due to different types of result objects

blueprint fuel-qa-python3-compatibility
Related-bug: #1556791

+++ fuelweb_test/tests/test_ha_one_controller.py
@@ -12,6 +12,8 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
+from __future__ import division
+
 import re
 
 from devops.helpers.helpers import wait
@@ -674,7 +676,7 @@ class NodeDiskSizes(TestBasic):
             for disk in node['meta']['disks']:
                 assert_equal(disk['size'], disk_size, 'Disk size')
 
-        hdd_size = "{} TB HDD".format(float(disk_size * 3 / (10 ** 9)) / 1000)
+        hdd_size = "{} TB HDD".format((disk_size * 3 / (10 ** 9)) / 1000)
         notifications = self.fuel_web.client.get_notifications()
         for node in nailgun_nodes:
             # assert /api/notifications
